requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3
python-magic-bin==0.4.14; platform_system == "Windows"
python-magic==0.4.27; platform_system != "Windows"